Fix a bug that will not correct show statemachine sprite after texture refreshing. Adding StateMachine.class case for GO.copy().

+++ mario/src/main/java/util/Time.java
@@ -7,9 +7,9 @@ package util;
 public class Time {
   public static long timeStarted = System.nanoTime();
 
-  public static double getTime() {
+  public static float getTime() {
     // The time is in nanosecond, we need to convert to second.
-    return ((System.nanoTime() - timeStarted) * 1E-9); //conversion
+    return (float) ((System.nanoTime() - timeStarted) * 1E-9); //conversion
   }
 
 }
